:truck: rename utility folder
