+++ app/utils/validation.py
@@ -1,0 +1,22 @@
+import re
+
+passwordError = '''Password is not strong enough: Requires at least: 1 lower case character, 1 upper case character, 1 digit, 1 special character and a length of 8 character'''
+
+def validate_image_file_type(content_type: str) -> bool:
+    regexp = "image/[^\?(.*)]"
+    return bool(re.search(regexp, content_type))
+
+def get_file_type(content_type: str):
+    regexp = "image/([^\/]+$)"
+    res = re.findall(regexp, "image/jpg")
+    if res != None:
+        return res[0]
+    return res
+
+def validate_password(password:str) -> bool:
+    regexp = "^(?=.*[a-z\æøå])(?=.*[A-Z\ÆØÅ])(?=.*\d)(?=.*[@$!%*?&])[A-Z\ÆØÅa-z\æøå\d@$!%*?&]{8,}$"
+    if not re.findall(regexp, password):
+        return False
+
+    return True
+
